Reorganized layout, callbacks moved to calbacks folder
